step 13

+++ lib/screens/notes_list_view.dart
@@ -2,16 +2,18 @@ import 'package:flutter/material.dart';
 import '../services/CRUD/note_services.dart';
 import '../utiles/delete_dialog.dart';
 
-typedef DeleteNoteCallback = void Function(DatabaseNotes note);
+typedef NoteCallback = void Function(DatabaseNotes note);
 
 class NotesListView extends StatelessWidget {
-  final DeleteNoteCallback onDeleteNote;
+  final NoteCallback onDeleteNote;
   final List<DatabaseNotes> notes;
+  final NoteCallback onTap;
 
   const NotesListView({
     Key? key,
     required this.onDeleteNote,
     required this.notes,
+    required this.onTap,
   }) : super(key: key);
 
   @override
@@ -21,6 +23,9 @@ class NotesListView extends StatelessWidget {
       itemBuilder: (context, index) {
         final note = notes[index];
         return ListTile(
+          onTap: () {
+            onTap(note);
+          },
           title: Text(
             note.text,
             maxLines: 1,
@@ -29,10 +34,10 @@ class NotesListView extends StatelessWidget {
           ),
           trailing: IconButton(
             onPressed: () async {
-             final shouldDelete=await showDeleteDialog(context);
-             if(shouldDelete){
-               onDeleteNote(note); // Call onDeleteNote callback with the note
-             }
+              final shouldDelete = await showDeleteDialog(context);
+              if (shouldDelete != null && shouldDelete) {
+                onDeleteNote(note); // Call onDeleteNote callback with the note
+              }
             },
             icon: Icon(Icons.delete),
           ),
